Rename config to game_config
