+++ tests/test_actions.py
@@ -4,7 +4,7 @@ from blackjack.hand import Hand
 from blackjack.deck import Deck
 from blackjack.deck import Card
 from blackjack.actions import continue_playing, twist, stick_or_twist, reveal_cards
-import config
+import game_config
 
 
 def test_twist(my_hand: Hand, card_deck: Deck):
@@ -91,17 +91,17 @@ def test_reveal_all_cards(mock_print, my_hand: Hand):
 @patch("builtins.input", lambda *args: "y")
 def test_continue_playing_yes():
     """Asserts that config.playing is true after continue_playing is called and the user input "y" """
-    config.playing = False
+    game_config.playing = False
     continue_playing()
-    assert config.playing is True
+    assert game_config.playing is True
 
 
 @patch("builtins.input", lambda *args: "n")
 def test_continue_playing_no():
     """Asserts that config.game_on is false after continue_playing is called and the user input "n" """
-    config.game_on = True
+    game_config.game_on = True
     continue_playing()
-    assert config.game_on is False
+    assert game_config.game_on is False
 
 
 @patch("builtins.input", side_effect=["x", "n"])
